dataAPP/requirements.txt
adbutils==2.8.0
apkutils2==1.0.0
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
cigam==0.0.3
click==8.1.7
colorama==0.4.6
deprecation==2.1.0
Flask==3.1.0
Flask-Cors==5.0.0
idna==3.10
ImageHash==4.3.1
itsdangerous==2.2.0
Jinja2==3.1.4
labelImg==1.8.6
lxml==5.3.0
MarkupSafe==3.0.2
MouseInfo==0.1.3
mss==9.0.2
numpy==2.2.0
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.3
patool==3.1.0
pillow==10.4.0
puremagic==1.28
py==1.11.0
PyAutoGUI==0.9.54
pyelftools==0.31
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pyserial==3.5
python-magic-bin==0.4.14
pytweening==1.2.0
pytz==2024.2
PyWavelets==1.8.0
requests==2.32.3
retry==0.9.2
scipy==1.14.1
screeninfo==0.8.1
setuptools==75.1.0
tzdata==2024.2
uiautomator2==3.2.5
urllib3==2.2.3
Werkzeug==3.1.3
wheel==0.44.0
xmltodict==0.14.0
